requirements.txt
2captcha-python==1.1.3
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
Brotli==1.0.9
capmonstercloudclient==1.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
discord==2.1.0
discord.py==2.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
multidict==6.0.4
outcome==1.2.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.7.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
undetected-chromedriver==3.2.1
urllib3==1.26.13
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
zstandard==0.19.0
